fix(ci): resolve lint failures blocking CI on main
Extract restore loop into restore-loop-executor.ts to satisfy the
300-line limit and remove unused imports flagged by ESLint.
