+++ packages/core/tests/unit/adapters/identity/caching-identity-resolver.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect, vi, beforeEach } from 'vitest';
+import { describe, it, expect, vi } from 'vitest';
 import type {
   IdentityRegistry,
   IdentityRegistryEntry,
